requirements.txt
GitPython==3.1.44
Jinja2==3.1.6
MarkupSafe==3.0.2
PyYAML==6.0.2
Pygments==2.19.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==6.0.0
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.9.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
distlib==0.3.9
docker-pycreds==0.4.0
dvc==3.60.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
huggingface-hub==0.33.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lightning-utilities==0.14.3
matplotlib-inline==0.1.7
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.4
mypy==1.16.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.2.4
omegaconf==2.3.0
pandas==2.3.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pyproject-api==1.9.1
pytest-cov==6.2.1
pytest==8.4.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pywin32==310
pyzmq==27.0.0
requests==2.32.4
ruff==0.11.12
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.30.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.6.0
tomli==2.2.1
torch==2.7.1
torchaudio==2.7.1
torchmetrics==1.7.3
torchvision==0.22.1
tornado==6.5.1
tox-gh-actions==3.3.0
tox==4.26.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.0
types-PyYAML==6.0.12.20250516
urllib3==2.4.0
virtualenv==20.31.2
wandb==0.20.1
wcwidth==0.2.13
yarl==1.20.1
